% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_user_name.R
\name{clean_user_name}
\alias{clean_user_name}
\title{Clean up RATSS user names}
\usage{
clean_user_name(entries)
}
\arguments{
\item{entries}{character vector}
}
\description{
Takes a character vector of  entries that
have a user name in the format "<digit/digits>_<single digit>"
somewhere in them, extracts only the user names and returns a vector
with those.
}
\details{
Note that the function can be used with a single entry with a user name
(i. e. a single element vector) as well as multiple entries. You need to make sure
that there aren't any digits immediately preceding or following the user name in
the entry string (e. g. "99102_2", with 102_2 being the username, doesn't work).
}
\examples{
foo <- c("Twin202_2", "313_5", "1_2", "1_1", "randomchars12_2fillout")
clean_user_name(foo)

}
\seealso{
\code{\link[stringr]{str_extract}}
}
