% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnr2bdate.R
\name{pnr2bdate}
\alias{pnr2bdate}
\title{Convert a personal number to birth date}
\usage{
pnr2bdate(pnr)
}
\arguments{
\item{pnr}{a character vector with a single personal number}
}
\description{
Helper function for \code{\link[ratss]{pnr_to_birth_date}}.
}
\details{
See the documentation for \code{\link[ratss]{pnr_to_birth_date}}.
}
\seealso{
\code{\link[ratss]{pnr_to_birth_date}}
}
