% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnr_to_birth_date.R
\name{pnr_to_birth_date}
\alias{pnr_to_birth_date}
\title{Convert a personal number to birth date}
\usage{
pnr_to_birth_date(pnrs)
}
\arguments{
\item{pnrs}{a character vector of personal numbers}
}
\description{
Function for converting Swedish personal numbers to birth dates.
}
\details{
Converts Swedish personal numbers (in the format "YYYYMMDD-XXXX",
or "YYMMDD-XXXX") to birth dates. If the personal number is in the format
"YYMMDD-XXXX", the function assumes that all YY values above 25 indicate
someone born in the 1900's, and all YY values below 25 indicate someone
born in the 2000's.
}
\examples{
x <- "19400830-1111"
pnr_to_birth_date(x)
y <- "090830-1111"
pnr_to_birth_date(y)
mult_bds <- c("19400830-1111", "19801212-0000", "20020202-2222", "000303-3333")
pnr_to_birth_date(mult_bds)

}
\seealso{
\code{\link[base]{sub}}
}
