% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat2usern.R
\name{rat2usern}
\alias{rat2usern}
\title{Convert a RATSS study ID to user name format}
\usage{
rat2usern(rawn)
}
\arguments{
\item{rawn}{a string/character vector with a single RATSS code}
}
\description{
Helper function for \code{\link[ratss]{ratid_to_user_name}}.
Converts a RATSS code/study ID to a user name.
}
\details{
See \code{\link[ratss]{ratid_to_user_name}}.
}
\examples{
x <- "11 0104 01 01 14"
rat2usern(x)
y <- "110104010114"
rat2usern(y)
}
\seealso{
\code{\link[ratss]{ratid_to_user_name}}
}
