% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratss.R
\docType{package}
\name{ratss}
\alias{ratss}
\alias{ratss-package}
\title{synratss}
\description{
This package was developed for use at the Center of
Neurodevelopmental Disorders at Karolinska Institutet (KIND). Its
applicability is highly narrow and specific. It has simple functions for
converting study ID's to user names, Swedish personal numbers to birth dates,
and finding and cleaning up RATSS user names.
}
